pyproject.toml
[tool.poetry]
name = "meteowebsite"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.1"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"
django-filter = "^21.1"
django-compression-middleware = "^0.4.1"
whitenoise = "^5.3.0"
hypercorn = "^0.13.2"
autopep8 = "^1.6.0"
python-dateutil = "^2.8.2"
psycopg2 = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
